[klibc] run_init() now returns an error message instead of calling die()
Calling a global function which can't return is pretty antisocial for
a librarized function.  It's possible to pass in a suitable die function
as an argument, which could contain either exit or longjmp depending on
the needs of the caller.  However, this is overkill for run_init(), which
really only just needs to return the error message, if applicable.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>