[klibc] dash: force signedness where needed to make cross-compile-safe
dash used to rely on mksyntax being run on the same architecture as
the final binary in order to determine signedness of "char".  This
doesn't work for cross-compiling.

Hi Peter:

Here is the patch.  I chose signed char instead of unsigned char as
the canonical form as that generates better/smaller code on x86.

I've verified that the assembly output is identical on x86 except
for a real bug fix in expand.c.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>