klibc: delete havesyscall.h during make clean
In the kernel KLIBCINC expands to a full path, whereas in klibc
it expands to a relative path.
So specify havesyscalls.h twice to cover both situations.
Note: kbuild will prefix relative paths with $(obj) when deleting.

Later we may use same type of definition for KLIBCINC
but until then we keep this workaround.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>