klibc/kbuild: use separate kbuild files for each klibc subdirectory
This fixes a long standing issue where it was not possible to
do "make usr/klibc/arch/x86_64/longjmp.o" in the kernel.

The principle is that all .o files to be part of klibc are listed
with klib-y. For each directory a klib.list file is made that specify
all .o file and the final AR then adds all .o files to create libc.a.

This patch introduce the infrastructure and converts x86_64 to use it.
And it breaks all other architectures...

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>