[klibc] Remove SA_RESTORER/sigreturn hack for i386
Remove the SA_RESTORER/sigreturn hack for i386.  On i386, the format
of the sigreturn stub varies depeding on if SA_SIGINFO is set or not;
the stock kernels were always OK (only some Fedora kernels were broken),
and the Fedora kernels that were broken can now be considered historic.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>