[klibc] [EXPAND] Do not expand tilde in parameter expansion within quotes
Previously code was added so that tilde expansion was carried out
parameter expansions within double quotes.  This change was made
with reference the behaviour of bash at the time.  Bash has since
be fixed so that this behaviour no longer occurs which is in line
with most other POSIX shells.

So this patch removes that behaviour in dash as well.

Test case:

	unset a
	echo "${a:-~root}"

Old result:

	/root

New result:

	~root

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: maximilian attems <max@stro.at>