[klibc] [PARSER] Recognise here-doc delimiters terminated by EOF
Previously dash required a <newline> character to be present in order for
a here-document delimiter to be detected.  Allowing EOF in the absence of
a <newline> to play the same purpose allows some intuitive scripts to
succeed.  POSIX seems to be silence on this so this should be OK.

Test case:

	eval 'cat <<- NOT
		test
	NOT'
	echo OK

Old result:

	test
	NOTOK

New result:

	test
	OK

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: maximilian attems <max@stro.at>