[klibc] mount: allow multiple fs on vfstype arg
Basically it fixes the argument parsing when running something like
"mount -t ext3,ext2...".

I've already submitted the attached patch to the Debian BTS:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=480994

It is not yet the real fix for "-t auto", that could be faked by
replacing auto with a long list of fs types or running "fstype".

Signed-off-by: maximilian attems <max@stro.at>