fstype: Fix ext4/ext4dev probing
Enhance fstype so it properly takes into account whether or not the
ext4 and/or ext4dev filesystems are present, and properly handles the
test_fs flag.  The old code also has some really buggy checks --- for
example, where it compared the set of supported ro_compat features
against the incompat feature bitmask:

	(sb->s_feature_incompat & __cpu_to_le32(EXT3_FEATURE_RO_COMPAT_SUPP)

I rewrote the ext4 checks so they are more easily understood and audited.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>