[klibc] Support for multiple devices specified on the root= cmdline.
In order to handle multiple root devices, such as root=/dev/hda1,/dev/sda1, we
introduce a new internal dev_t called Root_MULTI.  This device is returned if
name_to_dev sees a comma in the root line.

When handling multiple roots, we are careful to not support running linuxrc as
mixing multiple roots together with linuxrc doesn't make a lot of sense.

When we get to mounting and we have a Root_MULTI dev_t, we parse the comma
seperated tokens out of the cmdline argument and try each in succession,
stopping when we have a successful mount.

Signed-off-by: Mike Waychison <mikew@google.com>
Signed-off-by: maximilian attems <max@stro.at>