assert.h: Document that NDEBUG suppresses side effects
The C standard requires that defining NDEBUG suppresses not just the
faulting behavior of assert() but also any side effects of its
arguments.  This is implemented by checkin
c3cc6f14: [klibc] Fix assert when NDEBUG defined

Since this is also a common programming mistake when defining macros,
add a comment explaining why this is intentional.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>