[klibc] [BUILTIN] Use faccessat if available
Eric Blake suggested that we should use faccessat so that ACLs
and other corner cases are handled correctly.  This patch does
exactly that.

Note that faccessat doesn't handle ACLs when euid != uid, as
this case is currently implemented by glibc instead of the kernel,
using code similar to the existing dash test.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

[ config.h undef HAVE_FACCESSAT ]
Signed-off-by: maximilian attems <max@stro.at>