[klibc] sh4: syscalls fixes
Using the patch from bug #574834 klibc now builds, but does not work. It
assumes that the kernel returns the result of the syscalls in register
r3, while it behaves like a normal call, with the results in register
r0.

Also the pipe syscall needs a special handling as on some other
architectures, the structure passed in argument is not filled, instead
the two fd are returned in r0 and r1.

The patch below fixes the problem. It has been tested by creating an
initrd with initramfs and booting the system with it.

Closes: #578076

Signed-off-by: maximilian attems <max@stro.at>