[klibc] gzip: Cleanup unused variable
The variable is not used in our gzip copy.
In the gzip code it is used to strip of .exe of progname,
which makes no sense in a unixish environment.

Cleans up build with recent gcc 4.6 of:
usr/gzip/gzip.c: In function ‘main’:
usr/gzip/gzip.c:231:9: warning: variable ‘proglen’ set but not used [-Wunused-but-set-variable]

Signed-off-by: maximilian attems <max@stro.at>