[klibc] Use GCC definitions for size_t, ssize_t, ptrdiff_t
Fall back to (unsigned) long if klibc userspace applications are
compiled with something that is not GCC or close enough to also
define its __SIZE_TYPE__ and __PTRDIFF_TYPE__ macros.

Change the definition of ssize_t to be decoupled from ptrdiff_t
and be derived from __SIZE_TYPE__ instead, dropping the unsigned.

Signed-off-by: Thorsten Glaser <tg@mirbsd.org>
Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>