[klibc] syscalls: Fixup some of the -at syscall declarations
mknodat and mkdirat contain a spurious repeated parameter, linkat
is missing the final int flags parameter, symlinkat has the first
two parameters transposed, and fchmodat is missing the flags
parameter and is declared in the wrong header.

Also declarations are missing for utimensat and fchownat.

This patch fixes up these syscall declarations.

Originally-by: Neil Williams <codehelp@debian.org>
Originally-by: Anil Singhar <anil.singhar@linaro.org>
Signed-off-by: Steve Capper <steve.capper@linaro.org>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>